+++ CHANGES.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * keep CRS/Bounds for STAC Item is not all proj attributes are set in properties [#154](https://github.com/sentinel-hub/titiler-openeo/pull/154)
 
+## Added
+
+* Full User Defined Process Graph support including graph validation [#158](https://github.com/sentinel-hub/titiler-openeo/pull/158)
+
 ## [0.6.1] (2025-12-04)
 
 ## Fixed